changed sum to a float because it should be a float

+++ pagerank.c
@@ -57,7 +57,7 @@ float calculateWin(URL v, PRNode u, Graph web)
     //for i in v's outlinks: add inlink
     // actual sum loop
     Link curr = v->outLink;
-    int sum = 0;
+    float sum = 0;
     for (; curr != NULL; curr = curr->next) {
         sum = sum + curr->URLPointer->numInLinks;
     }
@@ -71,7 +71,7 @@ float calculateWout(URL v, PRNode u, Graph web)
     float top = u->nOutLinks;
     // Find v in graph.
     Link curr = v->outLink;
-    int sum = 0;
+    float sum = 0;
     // For every outlink of v, add its outlinks.
     for(; curr != NULL; curr = curr->next)
         sum = sum + curr->URLPointer->numOutLinks;
